+++ packages/christmas-elf-name-generator/example/remix.config.js
@@ -2,12 +2,10 @@
  * @type {import('@remix-run/dev/config').AppConfig}
  */
 module.exports = {
-  appDirectory: 'app',
   browserBuildDirectory: 'public/build',
+  ignoredRouteFiles: ['**/.*'],
   publicPath:
     process.env.NODE_ENV === 'development'
       ? '/build/'
-      : '/visma/christmas-elf-name-generator/build/',
-  serverBuildDirectory: 'build',
-  devServerPort: 8002,
+      : '/twoday/christmas-elf-name-generator/build/',
 };