+++ packages/cra-template-craco-super-template/template/src/api/index.js
@@ -2,4 +2,4 @@
 
 export * from './client';
 
-// https://www.npmjs.com/package/@visma/react-openapi-client-generator#mutations-and-updates
+// https://www.npmjs.com/package/@twoday/react-openapi-client-generator#mutations-and-updates