+++ packages/react-app-locale-utils/src/types/index.d.ts
@@ -1,4 +1,4 @@
-declare module '@visma/vite-plugin-date-fns-locale/dynamic-import-date-fns-locales' {
+declare module '@twoday/vite-plugin-date-fns-locale/dynamic-import-date-fns-locales' {
   interface T {
     [locale: string]: () => Promise<{ default: import('date-fns').Locale }>;
   }