+++ packages/react-app-locale-utils/src/useDateFnsLocale.ts
@@ -1,5 +1,5 @@
 import { create } from '@postinumero/use-async';
-import dateFnsLocaleDynamicImports from '@visma/vite-plugin-date-fns-locale/dynamic-import-date-fns-locales';
+import dateFnsLocaleDynamicImports from '@twoday/vite-plugin-date-fns-locale/dynamic-import-date-fns-locales';
 import useLocale from './useLocale.js';
 
 const [useImportDateFnsLocale] = create((locale: string) =>