+++ packages/react-intl-bundled-messages/src/types/index.d.ts
@@ -1,4 +1,4 @@
-declare module '@visma/vite-plugin-react-intl-bundled-messages/dynamic-import-messages' {
+declare module '@twoday/vite-plugin-react-intl-bundled-messages/dynamic-import-messages' {
   interface T {
     [locale: string]: () => Promise<{
       default: import('react-intl').IntlProvider['props']['messages'];