+++ packages/tailwindcss-nordic-cool/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-  vanillaTheme: require('./vanilla-theme.js'),
-  complementaryTheme: require('./complementary-theme.js'),
-};